Two Laters of Terrain, Granularity on PerlinNoise
Two layers of Terrain are now being loaded.
The Granilarity can be changed for PerlinNoise (in ChunkManager). Now it
needs to be changed so each Chunk has its own granularity

+++ LearnOpenGL/PerlinNoise.h
@@ -24,7 +24,7 @@ public:
 	GLfloat **  generate(GLint x, GLint y);
 
 	// Starting GLfloat impl
-	GLfloat generate(GLint x, GLint y, GLint z);
+	GLfloat generate(GLfloat x, GLfloat y, GLfloat z);
 
 	void setChunk(glm::vec3 *** gradientVec);
 	void setChunk(glm::vec2 ** gradientVec);
